Add many other inventory types
Not all of them are listed but they are the most common ones. It would be pretty hard
to list all of them, unless we find a way to do it without using the name. Then again,
some inventories have inconsistent behaviors with the middle-click so we can't just
blindly use it instead of the left-click in all of them