ASM-API: Initial
The assembly transformer api goal is to provide a flexible way to transform classes. The current one doesn't help the mod to support multiple versions. And, It has many duplicates to maintain and other issues. The new api takes the class node as input. And, changes the node by the transformers.