Dungeon Fix: Doesn't detect teammates with symbols
In dungeons, any teammate who has '_' in their name is not included in the map.
That's because the character '_' is being filtered using `TextUtils#keepScoreboardCharacters`

This commit solves the problem by adding `_` to the pattern.

Sometimes, the prefix of the health color has a repeat (e.g. `§a§a`).
This causes the pattern to be broken.

This commit solves the problem by ignoring the repeated part.