Dev Feature: Copy Block Data Mode
This feature provides the developer a way to copy the block data
With the tile entity of that block as NBT.

When the mode is on, Any right click and without any item on the player hand.
It copies the block hovered by the player's cursor as NBT to the clipboard.

The developer can toggle the mode using `/sba copyBlockData`

The mode is **disabled by default**.
It requires the **developer mode**.