Dungeons Fix: Reset isn't being detected
This issue has been caused since hypixel removed the server ID from players list.
And, the sidebar when the player is in a dungeon game.
This prevents the `DungeonUtils#reset` being invoked.
Which could results in a memory leak for the player.

This commit fixes the issue by detecting the change of servers in the chat message.
Which invokes `DungeonUtils#reset`.