Added an option to disable in other games
I apologize for anything I may have done incorrectly. This is my first time contributing to an open-source project :). Also a note: the toggle for showing in other games says "show in other games" rather than "alert in other games," but I figured that it would be alright, at least for the beta.