Fixed tests failing due to Locale changes
While i was developing a possible feature for the mod, i stumbled across a problem where one of the Tests was doing an AssertionError, and it was preventing me to run the client for testing, looking further into the issue, i've found out the original code used a DecimalFormat constructor which uses the default Locale on where the JVM is running (on my case, Brazil). So, i just changed the implementation by setting the Locale manually and using a factory method contained in NumberFormat class that solves the internationalization problem and formats the number according to the Pattern(since it uses US pattern, which matched with the one in the original code).

Doing that ended up fixing the issue and all tests ran without any problems and i was able to run the client again!