Add async localized strings loading - Strings are now loaded asynchronously on startup and when the player changes the language setting - Fix language button loading strings every single frame for every single button - Remove executor service shutdown for now
## Notes:
There's a side effect of the flag changing like .25s after the strings change since the local copy gets loaded first but the language setting is set after loading the online version. Can probably change this later. Also, the string loading system is kind of half part of the main data loading system and half its own thing rn. We have to decide if we can merge it into the main thing or if it has to be its own separate thing.